Rewrote clive.
Everything is new, the parser has been rewritten to be simpler but also
allow parsing of more complex things. It should work as expected every
time now.

Removed #flag and #switch which are now replaced by the generic #option,
this can take arguments and can not as well. Commands can also now take
arguments, and Arguments are a lot more powerful as they can be
constrained and coerced into different types.