Brought back #bool.
Boolean options are not a subclass of Option but just Options with an
option set (:boolean => true). The previous way of putting as: Boolean
then checking for a single argument was stupid and likely to break, not
to mention over complicated. This is easy.

opt :verbose, :as => Boolean  -->>   bool :verbose